+++ braindecode/functional/__init__.py
@@ -1,22 +1,28 @@
 from .functions import (
     _get_gaussian_kernel1d,
+    daubechies_filters,
     drop_path,
+    dwt_max_level,
     hilbert_freq,
     identity,
     plv_time,
     safe_log,
     square,
+    wavelet_decomposition,
 )
 from .initialization import glorot_weight_zero_bias, rescale_parameter
 
 __all__ = [
     "_get_gaussian_kernel1d",
+    "daubechies_filters",
     "drop_path",
+    "dwt_max_level",
     "hilbert_freq",
     "identity",
     "plv_time",
     "safe_log",
     "square",
+    "wavelet_decomposition",
     "glorot_weight_zero_bias",
     "rescale_parameter",
 ]